phoenix/llamaindex-py/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
llama-index-core>=0.12.0
llama-index-llms-anthropic>=0.6.0
chromadb>=0.6.0
pydantic>=2.0.0
arize-phoenix-otel>=0.6.0
openinference-instrumentation-llama-index>=0.1.0
arize-phoenix-client>=1.0.0
arize-phoenix-evals>=0.18.0
anthropic>=0.40.0
pandas>=2.0.0
